trading_assistant/requirements.txt
alpaca-py
python-dotenv
pyyaml
pandas
numpy